+++ main.py
@@ -3,16 +3,19 @@
 Run tracepaths on OpenFlow in the Control Plane
 """
 
+import ipaddress
 from datetime import datetime
 
+import requests
 from flask import jsonify, request
 from kytos.core import KytosNApp, log, rest
 from kytos.core.helpers import listen_to
-from napps.amlight.flow_stats.main import Main as FlowManager
 from napps.amlight.sdntrace_cp import settings
 from napps.amlight.sdntrace_cp.automate import Automate
-from napps.amlight.sdntrace_cp.utils import (convert_entries, find_endpoint,
-                                             prepare_json)
+from napps.amlight.sdntrace_cp.utils import (convert_entries,
+                                             convert_list_entries,
+                                             find_endpoint, prepare_json,
+                                             prepare_list_json)
 
 
 class Main(KytosNApp):
@@ -58,31 +61,50 @@ class Main(KytosNApp):
     def trace(self):
         """Trace a path."""
         entries = request.get_json()
-        result = self.tracepath(entries)
+        entries = convert_entries(entries)
+        stored_flows = self.get_stored_flows()
+        result = self.tracepath(entries, stored_flows)
         return jsonify(prepare_json(result))
 
-    def tracepath(self, entries):
+    @rest('/traces', methods=['PUT'])
+    def get_traces(self):
+        """For bulk requests."""
+        entries = request.get_json()
+        entries = convert_list_entries(entries)
+        stored_flows = self.get_stored_flows()
+        results = {}
+        list_ready = []
+        for entry in entries:
+            if (entry['dpid'], entry['in_port']) in list_ready:
+                continue
+            list_ready.append((entry['dpid'], entry['in_port']))
+            dpid = entry['dpid']
+            if dpid not in results:
+                results[dpid] = []
+            result = prepare_list_json(self.tracepath(entry, stored_flows))
+            results[dpid].append(result)
+        return jsonify(results)
+
+    def tracepath(self, entries, stored_flows):
         """Trace a path for a packet represented by entries."""
         self.last_id += 1
         trace_id = self.last_id
-        entries = convert_entries(entries)
         trace_result = []
         trace_type = 'starting'
-
         do_trace = True
         while do_trace:
             trace_step = {'in': {'dpid': entries['dpid'],
                                  'port': entries['in_port'],
                                  'time': str(datetime.now()),
                                  'type': trace_type}}
-            if 'vlan_vid' in entries:
-                trace_step['in'].update({'vlan': entries['vlan_vid'][-1]})
+            if 'dl_vlan' in entries:
+                trace_step['in'].update({'vlan': entries['dl_vlan'][-1]})
             switch = self.controller.get_switch_by_dpid(entries['dpid'])
-            result = self.trace_step(switch, entries)
+            result = self.trace_step(switch, entries, stored_flows)
             if result:
                 out = {'port': result['out_port']}
-                if 'vlan_vid' in result['entries']:
-                    out.update({'vlan': result['entries']['vlan_vid'][-1]})
+                if 'dl_vlan' in result['entries']:
+                    out.update({'vlan': result['entries']['dl_vlan'][-1]})
                 trace_step.update({
                     'out': out
                 })
@@ -115,12 +137,16 @@ class Main(KytosNApp):
                 return True
         return False
 
-    @staticmethod
-    def trace_step(switch, entries):
+    def trace_step(self, switch, entries, stored_flows):
         """Perform a trace step.
 
         Match the given fields against the switch's list of flows."""
-        flow, entries, port = FlowManager.match_and_apply(switch, entries)
+        flow, entries, port = self.match_and_apply(
+                                                    switch,
+                                                    entries,
+                                                    stored_flows
+                                                )
+
         if not flow or not port:
             return None
 
@@ -140,3 +166,101 @@ class Main(KytosNApp):
         # pylint: disable=unused-argument
         if settings.FIND_CIRCUITS_IN_FLOWS:
             self.automate.find_circuits()
+
+    @staticmethod
+    def get_stored_flows(dpids: list = None, state: str = None):
+        """Get stored flows from flow_manager napps."""
+        api_url = f'{settings.FLOW_MANAGER_URL}/stored_flows'
+        if dpids:
+            str_dpids = ''
+            for dpid in dpids:
+                str_dpids += f'&dpid={dpid}'
+            api_url += '/?'+str_dpids[1:]
+        if state:
+            char = '&' if dpids else '/?'
+            api_url += char+f'state={state}'
+        result = requests.get(api_url)
+        flows_from_manager = result.json()
+        return flows_from_manager
+
+    @classmethod
+    def do_match(cls, flow, args):
+        """Match a packet against this flow (OF1.3)."""
+        # pylint: disable=consider-using-dict-items
+        if ('match' not in flow['flow']) or (len(flow['flow']['match']) == 0):
+            return False
+        for name in flow['flow']['match']:
+            field_flow = flow['flow']['match'][name]
+            if name not in args:
+                return False
+            if name == 'dl_vlan':
+                field = args[name][-1]
+            else:
+                field = args[name]
+            if name not in ('ipv4_src', 'ipv4_dst', 'ipv6_src', 'ipv6_dst'):
+                if field_flow != field:
+                    return False
+            else:
+                packet_ip = int(ipaddress.ip_address(field))
+                ip_addr = flow['flow']['match'][name]
+                if packet_ip & ip_addr.netmask != ip_addr.address:
+                    return False
+        return flow
+
+    def match_flows(self, switch, args, stored_flows, many=True):
+        # pylint: disable=bad-staticmethod-argument
+        """
+        Match the packet in request against the stored flows from flow_manager.
+        Try the match with each flow, in other. If many is True, tries the
+        match with all flows, if False, tries until the first match.
+        :param args: packet data
+        :param many: Boolean, indicating whether to continue after matching the
+                first flow or not
+        :return: If many, the list of matched flows, or the matched flow
+        """
+        response = []
+        try:
+            for flow in stored_flows[switch.dpid]:
+                match = Main.do_match(flow, args)
+                if match:
+                    if many:
+                        response.append(match)
+                    else:
+                        response = match
+                        break
+        except AttributeError:
+            return None
+        if not many and isinstance(response, list):
+            return None
+        return response
+
+    # pylint: disable=redefined-outer-name
+    def match_and_apply(self, switch, args, stored_flows):
+        # pylint: disable=bad-staticmethod-argument
+        """Match flows and apply actions.
+        Match given packet (in args) against
+        the stored flows (from flow_manager) and,
+        if a match flow is found, apply its actions."""
+        flow = self.match_flows(switch, args, stored_flows, False)
+        port = None
+        actions = None
+        # pylint: disable=too-many-nested-blocks
+        if not flow or switch.ofp_version != '0x04':
+            return flow, args, port
+        actions = flow['flow']['actions']
+        for action in actions:
+            action_type = action['action_type']
+            if action_type == 'output':
+                port = action['port']
+            if action_type == 'push_vlan':
+                if 'dl_vlan' not in args:
+                    args['dl_vlan'] = []
+                args['dl_vlan'].append(0)
+            if action_type == 'pop_vlan':
+                if 'dl_vlan' in args:
+                    args['dl_vlan'].pop()
+                    if len(args['dl_vlan']) == 0:
+                        del args['dl_vlan']
+            if action_type == 'set_vlan':
+                args['dl_vlan'][-1] = action['vlan_id']
+        return flow, args, port