+++ settings.py
@@ -19,3 +19,5 @@ FIND_CIRCUITS_IN_FLOWS = False
 SDNTRACE_URL = 'http://localhost:8181/api/amlight/sdntrace/trace'
 
 SLACK_CHANNEL = 'of_notifications'
+
+FLOW_MANAGER_URL = 'http://localhost:8181/api/kytos/flow_manager/v2'