+++ utils.py
@@ -3,17 +3,6 @@
 from kytos.core import KytosEvent
 from napps.amlight.sdntrace_cp import settings
 
-TRANSLATE_NAMES = {
-    'dl_src': 'eth_src',
-    'dl_dst': 'eth_dst',
-    'dl_type': 'eth_type',
-    'dl_vlan': 'vlan_vid',
-    'nw_src': 'ip4_src',
-    'nw_dst': 'ip4_dst',
-    'nw_tos': 'ip_tos',
-    'nw_proto': 'ip_proto',
-}
-
 
 def convert_entries(entries):
     """ Transform entries dictionary in a plain dictionary suitable for
@@ -25,13 +14,19 @@ def convert_entries(entries):
     new_entries = {}
     for entry in entries['trace'].values():
         for field, value in entry.items():
-            if field in TRANSLATE_NAMES:
-                new_entries[TRANSLATE_NAMES[field]] = value
-            else:
-                new_entries[field] = value
-    if 'vlan_vid' in new_entries:
-        new_entries['vlan_vid'] = [new_entries['vlan_vid']]
+            new_entries[field] = value
+    if 'dl_vlan' in new_entries:
+        new_entries['dl_vlan'] = [new_entries['dl_vlan']]
     return new_entries
+
+
+def convert_list_entries(entries):
+    """ Transform a list of entries dictionary in a list
+    of plain dictionary suitable for matching
+    :param entries: list(dict)
+    :return: list(plain dict)
+    """
+    return [convert_entries(entry) for entry in entries]
 
 
 def find_endpoint(switch, port):
@@ -46,12 +41,17 @@ def find_endpoint(switch, port):
     return None
 
 
-def prepare_json(trace_result):
-    """Prepare return json for REST call."""
+def prepare_list_json(trace_result):
+    """Prepare return list of json for REST call."""
     result = []
     for trace_step in trace_result:
         result.append(trace_step['in'])
-    return {'result': result}
+    return result
+
+
+def prepare_json(trace_result):
+    """Prepare return json for REST call."""
+    return {'result': prepare_list_json(trace_result)}
 
 
 def format_result(trace):