+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ All notable changes to the sdntrace_cp NApp will be documented in this file.
 Added
 =====
 - Added ``apscheduler`` library to handle job scheduling
+- Added `PUT /traces` endpoint for bulk requests
 
 Changed
 =======